Update codebase for better code readability and data structuring
This commit involves various updates to improve the readability of the code and to enhance the structuring of data. Changes include renaming of a data structure in the "connection.go" file for better clarity, addition of extra fields in the "config.go" file for more comprehensive profiling, and reordering of output in the "profile.go" file for improved readability. Error handling has also been improved across different files.


Former-commit-id: f76e3d2969062d25de4357d30463e6fbb816021c [formerly 5efa5da7d8642ef15c3f51d274ea569ff4e09f0a]
Former-commit-id: 162f682ce0870e89260d1a9c3789f93ea475093c