Remove nip19 package and refactor associated inputs
This commit removes the nip19 package entirely. The refactoring includes updating nip19 related input method calls in other modules. These changes ensure the program maintains functionality despite the removal of the nip19 package. Primarily, the references were switched from the 'go-nostr' package to the 'nostr' package.

the reason for doing this is the retarded github actions are failing the build because of a nonexistent lack of the nip19 package that was deleted in this commit for this reason, fingers crossed this time it works


Former-commit-id: f96e563d17033e895ccf7c96319fe400c433c8c3 [formerly 901092cc2940459bee5467bf42e90ca01aef0b2e]
Former-commit-id: a9f06ebd07f489e28bd020d5507f651c2b2c88ba