Update cmd/postr package and add Get function
Updated several function definitions in the cmd/postr package for better clarity and functionality. Most notably, implemented a new Get function in the config module which enables retrieval of one or more events based on provided ID hashes. This function is designed to work with the newly added 'get' command line interface, which supports JSON output and accompanies Event printing enhancements in the config module.


Former-commit-id: 6218bb25d436bbfa23e5a06fb573062a5438d23c [formerly a2fbf2dd86a16f884f0699382c9306266f900071]
Former-commit-id: 711e664bb860c57a28a3ac9de7495539fa224fb4