Add NIP-11 fetch function and related tests
The changes introduced include the addition of a fetch function to retrieve NIP-11 info and related tests to ensure correctness. The normalization of URLs to start with http:// or https:// has been implemented, and a default timeout of 7 seconds has been set if no other timeout is provided. The `AddSupportedNIP` function now includes binary search to improve performance.