Add NIP-11 fetch function and related tests
The changes introduced include the addition of a fetch function to retrieve NIP-11 info and related tests to ensure correctness. The normalization of URLs to start with http:// or https:// has been implemented, and a default timeout of 7 seconds has been set if no other timeout is provided. The `AddSupportedNIP` function now includes binary search to improve performance.


Former-commit-id: 4f4aada5905dc0cd3b2feb4512ce429e9cf3c0d9 [formerly 941842036586e9eeb5300eb70bb664aa6e511ae1]
Former-commit-id: 7a822816deaf97aae307dcfbe4c69c95f44020d0