Refactor envelope identification and processing workflow
This update refactors the envelope handling process. The changes mainly involve streamlining the envelope identification process in the 'Identify' function and altering the return type of 'match' from labels.T to a string. Furthermore, logic contained in the 'ProcessEnvelope' function has been removed and replaced with a call to 'Identify' from the sentinel package. This promotes a cleaner and more maintainable code structure.