Refactor envelope identification and processing workflow
This update refactors the envelope handling process. The changes mainly involve streamlining the envelope identification process in the 'Identify' function and altering the return type of 'match' from labels.T to a string. Furthermore, logic contained in the 'ProcessEnvelope' function has been removed and replaced with a call to 'Identify' from the sentinel package. This promotes a cleaner and more maintainable code structure.


Former-commit-id: 8b080c406acdfde2d75877f25a99989e3ef5b83e [formerly 987bdd8dc1b9b2d6c07832e2ae525faa8d055912]
Former-commit-id: 0a786614e1023dbbd795453e8496bfb2305268f6