Refactor envelope identification and reading in sentinel package
This commit streamlines the envelope identification process by changing the return type of 'match' from labels.T to a string. It removes redundant code from 'ProcessEnvelope' function and replaces it with a call to 'Read' function in sentinel. Furthermore, it optimizes the process of envelope unmarshaling and error handling. These changes contribute to cleaner, more manageable, and efficient code.