Refactor envelope identification and reading in sentinel package
This commit streamlines the envelope identification process by changing the return type of 'match' from labels.T to a string. It removes redundant code from 'ProcessEnvelope' function and replaces it with a call to 'Read' function in sentinel. Furthermore, it optimizes the process of envelope unmarshaling and error handling. These changes contribute to cleaner, more manageable, and efficient code.


Former-commit-id: 2cf255298a07536bc5d8251779b41a32d390258d [formerly a1b66627310e45b99787c230cadc7dcd52645c45]
Former-commit-id: f17d4f4ddbbb6ba3c0ff503232dc0150381ce0ea