Refactor Relay and Subscription, extract common methods to interface
The RelayConnect method in the Relay is refactored to Connect and MustRelayConnect is refactored to MustConnect. The Relay struct is updated to include an interface with methods rebased into it. The various Subscription method calls are altered to match this new interface. This allows the subscription type to be fully isolated from the relay


Former-commit-id: 81b4b8bd5947f22cc8a744c74916b22194c8ca91 [formerly 75af5324b872873a0290c53a92498a50b73adc73]
Former-commit-id: aeebc58fe3f2812e3fdb6846d79a1a3c5fc11fb5