Replace WriteJSON with WriteEnvelope in various files
This commit changes the function WriteJSON to WriteEnvelope in several files to improve consistency and readability. This updated function will still write an object but now does so as an envelope instead of JSON, which simplifies the code and eliminates redundancy. The commit also includes minor updates and fixes spread across multiple files, primarily addressing issues related to the handling of websocket communications.


Former-commit-id: b0af78d60c93c30c2d445f502185743785812115 [formerly 13e898adf52fec131212591d5c7922b362c89e59]
Former-commit-id: 7ce678fc9e81ed134cb79cc2233e633293baac71