Update NIP handling and tests
This change updates the handling of NIPs in the codebase. It introduces a structured definition for NIPs and their associated details in `types.go`. Correspondingly, it modifies the test cases to reflect this update. Furthermore, it adjusts the treatment of NIPs in `main.go`, with a clearer, more extensive representation of various NIPs.


Former-commit-id: 48d793dbe73711ef15ab6fdb00d60f9df24a453c [formerly 99c3b29942aac033e781b51db6b0b0b31ab0f06d]
Former-commit-id: ed08fbaf76b099c370da22f9ea6f2c5fd9f4a3ee