Reject old events and reduce envelope size
The commit adds a new functionality to reject events that are dated over an hour in the past, primarily to avoid the processing of outdated or irrelevant events. Also, it modifies the envelope processing function to reduce the number of bytes copied into the `trunc` slice, limiting it to 512 instead of 1024. This change is to optimize the handling of larger envelopes.


Former-commit-id: 5456e28943593dd11b444777c2d9fea7a1bbdd4b [formerly 6be7203c6cfbd132ea56080ee40fddb14fc0ff9f]
Former-commit-id: f7acb36cdbb0f6094bfc29d266d3bd2a104adc97