Add IC implementation and integrate it with main.go
A new package 'IC' was added to the codebase, which currently is a stub that calls all the methods of the 'badger.Backend'. This serves as groundwork for later inclusion of ICP storage driver functionality. Changes have been made in main.go to initiate the IC backend that wraps the existing 'badger.Backend' with its path and corresponding settings.