Add IC implementation and integrate it with main.go
A new package 'IC' was added to the codebase, which currently is a stub that calls all the methods of the 'badger.Backend'. This serves as groundwork for later inclusion of ICP storage driver functionality. Changes have been made in main.go to initiate the IC backend that wraps the existing 'badger.Backend' with its path and corresponding settings.


Former-commit-id: b2b38eb22bb86cbf4fc4c912de216ac816c46192 [formerly bd9274ed82aa49cf8b9db309fb71e0c2f725fde2]
Former-commit-id: 867f6098be1ca30e2c4c9454a12c3f911805ba9b