Implement whitelist and optimize message handling
The update introduces a whitelist feature to permit specific IP addresses for access and optimizes the handling of messages with better checks and adjustments. Most notably, it includes efficient handling of message processing, automatic limit setting for results when no limit is given and better management of query events. It also enhances the logging system providing more detailed information.


Former-commit-id: ec2dc463c8de47233e52e9637bbcf281015e6d67 [formerly 37c1199571d7042f45b8420184df73741ee8e049]
Former-commit-id: 4855c88d8f2d6a060072e79cbda61056968a811a